+++ engine/src/Core/Renderer/OpenGL/OpenGLIndexBuffer.cpp
@@ -8,7 +8,7 @@ namespace Turbo
         glCreateBuffers(1, &m_rendererId);
         glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, m_rendererId);
         glBufferData(GL_ELEMENT_ARRAY_BUFFER, indices.size() * sizeof(std::uint32_t), indices.data(), GL_STATIC_DRAW);
-        m_count = indices.size();
+        m_count = static_cast<std::uint32_t>(indices.size());
     }
 
     IndexBuffer<RenderingApi::OpenGL>::~IndexBuffer() { glDeleteBuffers(1, &m_rendererId); }
